+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 **v0.9.2**:
-- added support for governmental and sovereign clouds via new setting `powerbi.authenticationProvider`
+- added support for governmental and sovereign clouds via new setting `powerbi.cloud`
 - minor improvements for tooltips in treeviews
 
 **v0.9.1**: